pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bookverse-case"
version = "0.1.0"
description = "FastAPI service for the Smart Search in Books case"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115,<1.0",
    "uvicorn[standard]>=0.30,<1.0",
    "jinja2>=3.1,<4.0",
    "python-multipart>=0.0.9,<1.0",
    "python-dotenv>=1.0,<2.0",
    "httpx>=0.27,<1.0",
    "numpy>=1.26,<3.0",
    "faiss-cpu>=1.8,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2,<9.0",
]

[tool.setuptools.packages.find]
include = ["bookverse*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
